modify testList.cpp and add testConnection.cpp testHyperLogLog.cpp
