+++ test/testList.cpp
@@ -516,193 +516,6 @@ void TestBrpoplpush( void )
 	}
 }
 
-void TestPing( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		string value;
-		bool flag = redis.ping(value);
-		cout << flag << endl;
-		cout << value << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestQuit( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		bool flag = redis.quit();
-		cout << flag << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestEcho( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		string value;
-		bool flag = redis.echo("hello world", value);
-		cout << flag << endl;
-		cout << value << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestAuth( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		string value;
-		bool flag = redis.auth("1234");
-		cout << flag << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestSelect( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		string value;
-		bool flag = redis.select(1);
-		cout << flag << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestPfadd( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		CRedisClient::VecString element;
-		element.push_back("a");
-		element.push_back("b");
-		element.push_back("c");
-		string key("key1");
-		uint64_t count = redis.pfadd(key, element);
-		cout << count << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestPfcount( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		CRedisClient::VecString element;
-		element.push_back("a");
-		element.push_back("b");
-		element.push_back("c");
-		string key1("key1");
-		redis.pfadd(key1, element);
-
-		CRedisClient::VecString element2;
-		element2.push_back("a");
-		element2.push_back("d");
-		string key2("key2");
-		redis.pfadd(key2, element2);
-
-		CRedisClient::VecString keys;
-		keys.push_back(key1);
-		keys.push_back(key2);
-		uint64_t num = redis.pfcount(keys);
-		cout << num << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
-void TestPfmerge( void )
-{
-	try
-	{
-		CRedisClient redis;
-		redis.connect("127.0.0.1", 6379);
-
-		CRedisClient::VecString element;
-		element.push_back("a");
-		element.push_back("b");
-		element.push_back("c");
-		string key1("key1");
-		redis.pfadd(key1, element);
-
-		CRedisClient::VecString element2;
-		element2.push_back("a");
-		element2.push_back("d");
-		string key2("key2");
-		redis.pfadd(key2, element2);
-
-		CRedisClient::VecString keys;
-		keys.push_back(key1);
-		keys.push_back(key2);
-		bool flag = redis.pfmerge("test", keys);
-		cout << flag << endl;
-	} catch( RdException& e )
-	{
-		std::cout << "Redis exception:" << e.what() << std::endl;
-	} catch( Poco::Exception& e )
-	{
-		std::cout << "Poco_exception:" << e.what() << std::endl;
-	}
-}
-
 void Lmain( void )
 {
 	TestLpush();
@@ -722,14 +535,6 @@ void Lmain( void )
 	TestBlpop();
 	TestBrpop();
 	TestBrpoplpush();
-	TestPing();
-	TestQuit();
-	TestEcho();
-	TestAuth();
-	TestSelect();
-	TestPfadd();
-	TestPfcount();
-	TestPfmerge();
     return;
 }
